lab/sec4ml/GDALR-An-Efficient-Model-Duplication-Attack-on-Black-Box-Machine-Learning-Models/requirements.txt
numpy==1.15.4
Keras==2.2.4
matplotlib==3.0.2
pandas==0.22.0
torch==1.0.0
Pillow==5.4.1
scikit_learn==0.20.2
